+++ shiny/session/_session.py
@@ -24,7 +24,6 @@ from typing import (
     Callable,
     Iterable,
     Optional,
-    TypeVar,
     Union,
     cast,
     overload,
@@ -49,11 +48,9 @@ from ..http_staticfiles import FileResponse
 from ..input_handler import input_handlers
 from ..reactive import Effect_, Value, effect, flush, isolate
 from ..reactive._core import lock, on_flushed
-from ..render.transformer import OutputRenderer
+from ..render.renderer import Jsonifiable, RendererBase, RendererBaseT
 from ..types import SafeException, SilentCancelOutputException, SilentException
 from ._utils import RenderedDeps, read_thunk_opt, session_context
-
-OT = TypeVar("OT")
 
 
 class ConnectionState(enum.Enum):
@@ -951,6 +948,8 @@ class Inputs:
 # ======================================================================================
 # Outputs
 # ======================================================================================
+
+
 class Outputs:
     """
     A class representing Shiny output definitions.
@@ -969,7 +968,7 @@ class Outputs:
         self._suspend_when_hidden = suspend_when_hidden
 
     @overload
-    def __call__(self, renderer_fn: OutputRenderer[OT]) -> OutputRenderer[OT]:
+    def __call__(self, renderer: RendererBaseT) -> RendererBaseT:
         ...
 
     @overload
@@ -979,32 +978,31 @@ class Outputs:
         id: Optional[str] = None,
         suspend_when_hidden: bool = True,
         priority: int = 0,
-    ) -> Callable[[OutputRenderer[OT]], OutputRenderer[OT]]:
+    ) -> Callable[[RendererBaseT], RendererBaseT]:
         ...
 
     def __call__(
         self,
-        renderer_fn: Optional[OutputRenderer[OT]] = None,
+        renderer: Optional[RendererBaseT] = None,
         *,
         id: Optional[str] = None,
         suspend_when_hidden: bool = True,
         priority: int = 0,
-    ) -> OutputRenderer[OT] | Callable[[OutputRenderer[OT]], OutputRenderer[OT]]:
-        def set_renderer(renderer_fn: OutputRenderer[OT]) -> OutputRenderer[OT]:
-            if hasattr(renderer_fn, "on_register"):
-                renderer_fn.on_register()
-
-            # Get the (possibly namespaced) output id
-            output_name = self._ns(id or renderer_fn.__name__)
-
-            if not isinstance(renderer_fn, OutputRenderer):
+    ) -> RendererBaseT | Callable[[RendererBaseT], RendererBaseT]:
+        def set_renderer(renderer: RendererBaseT) -> RendererBaseT:
+            if not isinstance(renderer, RendererBase):
                 raise TypeError(
                     "`@output` must be applied to a `@render.xx` function.\n"
                     + "In other words, `@output` must be above `@render.xx`."
                 )
 
-            # renderer_fn is a Renderer object. Give it a bit of metadata.
-            renderer_fn._set_metadata(self._session, output_name)
+            # Get the (possibly namespaced) output id
+            output_name = self._ns(id or renderer.__name__)
+
+            # renderer is a Renderer object. Give it a bit of metadata.
+            renderer._set_output_metadata(output_name=output_name)
+
+            renderer._on_register()
 
             self.remove(output_name)
 
@@ -1019,12 +1017,9 @@ class Outputs:
                     {"recalculating": {"name": output_name, "status": "recalculating"}}
                 )
 
-                message: dict[str, Optional[OT]] = {}
+                message: dict[str, Jsonifiable] = {}
                 try:
-                    if _utils.is_async_callable(renderer_fn):
-                        message[output_name] = await renderer_fn()
-                    else:
-                        message[output_name] = renderer_fn()
+                    message[output_name] = await renderer.render()
                 except SilentCancelOutputException:
                     return
                 except SilentException:
@@ -1063,12 +1058,12 @@ class Outputs:
 
             self._effects[output_name] = output_obs
 
-            return renderer_fn
+            return renderer
 
-        if renderer_fn is None:
+        if renderer is None:
             return set_renderer
         else:
-            return set_renderer(renderer_fn)
+            return set_renderer(renderer)
 
     def remove(self, id: Id) -> None:
         output_name = self._ns(id)