+++ shiny/templates/package-templates/js-react/srcts/index.tsx
@@ -10,11 +10,12 @@ import {
 // into the root of the webcomponent.
 makeReactInput({
   name: "custom-component-input",
+  selector: "custom-component-input",
   initialValue: "#fff",
-  renderComp: ({ initialValue, onNewValue }) => (
+  renderComp: ({ initialValue, updateValue }) => (
     <ColorPickerReact
       initialValue={initialValue}
-      onNewValue={(color) => onNewValue(color)}
+      updateValue={(color) => updateValue(color)}
     />
   ),
 });
@@ -22,10 +23,10 @@ makeReactInput({
 // Color Picker React component
 function ColorPickerReact({
   initialValue,
-  onNewValue,
+  updateValue,
 }: {
   initialValue: string;
-  onNewValue: (x: string) => void;
+  updateValue: (x: string) => void;
 }) {
   const [currentColor, setCurrentColor] = React.useState(initialValue);
 
@@ -34,7 +35,7 @@ function ColorPickerReact({
       color={currentColor}
       onChange={(color) => {
         setCurrentColor(color.hex);
-        onNewValue(color.hex);
+        updateValue(color.hex);
       }}
     />
   );
@@ -42,6 +43,7 @@ function ColorPickerReact({
 
 makeReactOutput<{ value: string }>({
   name: "custom-component-output",
+  selector: "custom-component-output",
   renderComp: ({ value }) => (
     <div
       style={{